moved multifiledownloader to `files`, added changelog
