+++ context/app/static/js/components/detailPage/files/MultiFileDownloader/index.ts
@@ -1,0 +1,3 @@
+import MultiFileDownloader from './MultiFileDownloader';
+
+export default MultiFileDownloader;
